pcsx2 MMI: u16 -> u32 for src parameter of PMFHL_CLAMP.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@3696 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/MMI.cpp
@@ -171,11 +171,11 @@ void PLZCW() {
 	_PLZCW (1);
 }
 
-__fi void PMFHL_CLAMP(u16 dst, u16 src)
+__fi void PMFHL_CLAMP(u16 dst, u32 src)
 {
     if ((int)src > (int)0x00007fff)
 	    dst = 0x7fff;
-    else if ((int)src < (int)0xffff8000)  // Ints only go up to 0x7FFFFFFF. Something's not right here. --arcum42
+    else if ((int)src < (int)0xffff8000)
 	    dst = 0x8000;
     else
 	    dst = (u16)src;
